+++ proto/client-api/src/lib.rs
@@ -7,6 +7,19 @@
 //! - `medea`: Enables [`Deserialize`] implementation for [`Command`]s, and
 //! [`Serialize`] implementation for [`Event`]s.
 //! - `extended-stats`: Enables unused RTC Stats DTOs.
+//!
+//! # Contribution guide
+//!
+//! Avoid using 64 bit types. Jason uses [wasm-bindgen] to interop with JS,
+//! and exposing 64 bit types to JS will make [wasm-bindgen] to use
+//! [BigInt64Array][2] in its JS glue, which is not implemented or was
+//! implemented too recently in some UA's.
+//!
+//! So its better to keep protocol 64-bit-types-clean to avoid breaking things
+//! by accident.
+//!
+//! [wasm-bindgen]: https://github.com/rustwasm/wasm-bindgen
+//! [2]: https://tinyurl.com/y8bacb93
 
 pub mod stats;
 
@@ -25,7 +38,7 @@ use self::stats::RtcStat;
 )]
 #[cfg_attr(feature = "jason", derive(Serialize))]
 #[derive(Clone, Copy, Display)]
-pub struct PeerId(pub u64);
+pub struct PeerId(pub u32);
 
 /// ID of `MediaTrack`.
 #[cfg_attr(
@@ -34,7 +47,7 @@ pub struct PeerId(pub u64);
 )]
 #[cfg_attr(feature = "jason", derive(Serialize))]
 #[derive(Clone, Copy, Display)]
-pub struct TrackId(pub u64);
+pub struct TrackId(pub u32);
 
 /// Value that is able to be incremented by `1`.
 #[cfg(feature = "medea")]
@@ -71,7 +84,7 @@ impl_incrementable!(TrackId);
 pub enum ServerMsg {
     /// `ping` message that `Media Server` is expected to send to `Client`
     /// periodically for probing its aliveness.
-    Ping(u64),
+    Ping(u32),
 
     /// `Media Server` notifies `Client` about happened facts and it reacts on
     /// them to reach the proper state.
@@ -89,12 +102,12 @@ pub struct RpcSettings {
     /// doesn't receive [`ClientMsg::Pong`].
     ///
     /// Unit: millisecond.
-    pub idle_timeout_ms: u64,
+    pub idle_timeout_ms: u32,
 
     /// Interval that `Media Server` sends [`ServerMsg::Ping`] with.
     ///
     /// Unit: millisecond.
-    pub ping_interval_ms: u64,
+    pub ping_interval_ms: u32,
 }
 
 #[cfg_attr(test, derive(PartialEq))]
@@ -103,7 +116,7 @@ pub struct RpcSettings {
 pub enum ClientMsg {
     /// `pong` message that `Client` answers with to `Media Server` in response
     /// to received [`ServerMsg::Ping`].
-    Pong(u64),
+    Pong(u32),
 
     /// Request of `Client` to change the state on `Media Server`.
     Command(Command),
@@ -440,6 +453,7 @@ impl Serialize for ClientMsg {
 }
 
 #[cfg(feature = "medea")]
+#[allow(clippy::cast_possible_truncation)]
 impl<'de> Deserialize<'de> for ClientMsg {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
@@ -463,7 +477,7 @@ impl<'de> Deserialize<'de> for ClientMsg {
                 ))
             })?;
 
-            Ok(Self::Pong(n))
+            Ok(Self::Pong(n as u32))
         } else {
             let command =
                 serde_json::from_value::<Command>(ev).map_err(|e| {
@@ -500,6 +514,7 @@ impl Serialize for ServerMsg {
 }
 
 #[cfg(feature = "jason")]
+#[allow(clippy::cast_possible_truncation)]
 impl<'de> Deserialize<'de> for ServerMsg {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
@@ -523,7 +538,7 @@ impl<'de> Deserialize<'de> for ServerMsg {
                 ))
             })?;
 
-            Ok(Self::Ping(n))
+            Ok(Self::Ping(n as u32))
         } else {
             let msg = serde_json::from_value::<Event>(ev.clone())
                 .map(Self::Event)
